fix: revert redundant (int) casts where gfrv() already validates

+++ aggregate_graphs.php
@@ -337,7 +337,7 @@ function form_save() : void {
 			// generate a new sequence if needed
 			if (ierv('sequence')) {
 				$sequence = gfrv('sequence');
-				srv('sequence', get_sequence((int) $sequence, 'sequence', 'graph_templates_item', 'local_graph_id=' . (int) grv('local_graph_id')));
+				srv('sequence', get_sequence($sequence, 'sequence', 'graph_templates_item', 'local_graph_id=' . grv('local_graph_id')));
 			}
 			$save['id']                           = gfrv('graph_template_item_id');
 			$save['graph_template_id']            = gfrv('graph_template_id');
